Make the RPCClientFactory configurable
