+++ Assets/LoomSDK/IRPCClient.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: ca8350c8bbc617544b0e7acfab4294e4
+timeCreated: 1526557256
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
